--- shenyu-sdk-starter/src/main/java/org/apache/shenyu/sdk/starter/core/EnableShenyuClients.java
@@ -1,70 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.sdk.starter.core;
-
-import org.springframework.context.annotation.Import;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-/**
- * Scans for interfaces that declare they are Shenyu clients (via
- * {@link ShenyuClient} <code>@ShenyuClient</code>).
- * Configures component scanning directives for use with
- * {@link org.springframework.context.annotation.Configuration}
- * <code>@Configuration</code> classes.
- */
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-@Documented
-@Import(ShenyuClientsRegistrar.class)
-public @interface EnableShenyuClients {
-
-    /**
-     * Alias for the {@link #basePackages()} attribute. Allows for more concise annotation
-     * declarations e.g.: {@code @ComponentScan("org.my.pkg")} instead of
-     * {@code @ComponentScan(basePackages="org.my.pkg")}.
-     * @return the array of 'basePackages'.
-     */
-    String[] value() default {};
-
-    /**
-     * Base packages to scan for annotated components.
-     *
-     * <p>{@link #value()} is an alias for (and mutually exclusive with) this attribute.
-     *
-     * <p>Use {@link #basePackageClasses()} for a type-safe alternative to String-based
-     * package names.
-     * @return the array of 'basePackages'.
-     */
-    String[] basePackages() default {};
-
-    /**
-     * Type-safe alternative to {@link #basePackages()} for specifying the packages to
-     * scan for annotated components. The package of each class specified will be scanned.
-     *
-     * <p>Consider creating a special no-op marker class or interface in each package that
-     * serves no purpose other than being referenced by this attribute.
-     * @return the array of 'basePackageClasses'.
-     */
-    Class<?>[] basePackageClasses() default {};
-
-}
