--- shenyu-sdk-starter/src/main/java/org/apache/shenyu/sdk/starter/core/factory/AnnotatedParameterProcessor.java
@@ -1,55 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.sdk.starter.core.factory;
-
-import org.apache.shenyu.sdk.core.common.RequestTemplate;
-
-import java.lang.annotation.Annotation;
-
-/**
- * Shenyu contract method parameter processor.
- */
-public interface AnnotatedParameterProcessor {
-
-    /**
-     * Retrieves the processor supported annotation type.
-     * @return the annotation type
-     */
-    Class<? extends Annotation> getAnnotationType();
-
-    /**
-     * order.
-     *
-     * @return {@link int}
-     */
-    default int order() {
-        return Integer.MIN_VALUE;
-    }
-
-
-    /**
-     * Process the annotated parameter.
-     *
-     * @param requestTemplate requestTemplate
-     * @param annotation annotation
-     * @param arg arg
-     * @return {@link boolean}
-     */
-    boolean processArgument(RequestTemplate requestTemplate, Annotation annotation, Object arg);
-
-}
