+++ shenyu-sdk/pom.xml
@@ -27,17 +27,16 @@
 
     <artifactId>shenyu-sdk</artifactId>
 
+    <packaging>pom</packaging>
+    <modules>
+        <module>shenyu-sdk-core</module>
+        <module>shenyu-sdk-spring</module>
+    </modules>
+
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
-    <dependencies>
-        <dependency>
-            <groupId>org.apache.shenyu</groupId>
-            <artifactId>shenyu-spi</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
 </project>
